+++ inst/shiny/ui_files/ui_1_loadModels.R
@@ -246,14 +246,16 @@ ui.loadModels <- function() {
               condition = "input.model_select_action == 1",
               box(
                 title = "Interactive Preview", status = "primary", solidHeader = TRUE,  width = 12, collapsible = TRUE,
-                shinycssloaders::withSpinner(leafletOutput("model_preview_interactive_plot", height = 500), type = 1)
+                # shinycssloaders::withSpinner(leafletOutput("model_preview_interactive_plot", height = 500), type = 1)
+                leafletOutput("model_preview_interactive_plot", height = 500)
               )
             ),
             conditionalPanel(
               condition = "input.model_select_action == 2",
               box(
                 title = "Static Preview", status = "primary", solidHeader = TRUE,  width = 12, collapsible = TRUE,
-                shinycssloaders::withSpinner(plotOutput("model_preview_plot", height = 500), type = 1)
+                # shinycssloaders::withSpinner(plotOutput("model_preview_plot", height = 500), type = 1)
+                plotOutput("model_preview_plot", height = 500)
               )
             )
           )